fixed redundant type cast
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/cayenne/main/trunk@402398 13f79535-47bb-0310-9956-ffa450edef68

+++ cayenne/src/cayenne/java/org/objectstyle/cayenne/access/DefaultResultIterator.java
@@ -156,9 +156,9 @@ public class DefaultResultIterator implements ResultIterator {
         this.fetchLimit = fetchLimit;
 
         this.mapCapacity =
-            (int) Math.ceil(((double) descriptor.getNames().length) / 0.75);
+            (int) Math.ceil((descriptor.getNames().length) / 0.75);
         this.idMapCapacity =
-            (int) Math.ceil(((double) descriptor.getIdIndexes().length) / 0.75);
+            (int) Math.ceil((descriptor.getIdIndexes().length) / 0.75);
 
         checkNextRow();
     }
